Move DPO dataset configs to _experimental/ directory.
